+++ packages/react/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - **Combobox**: Add `ComboboxEmpty` component to display content when the combobox has no items
 
+- **Listbox**: Add `ListboxEmpty` component to display content when the listbox has no items
+
 ## [5.21.0] - 2025-08-24
 
 ### Added